Parallel-Query-Retrival/pyproject.toml
[project]
name = "parallel-query-retrival"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.9",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
]
